requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
av==16.0.1
backoff==2.2.1
bce-python-sdk==0.9.42
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.15
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cpm-kernels==1.0.11
crcmod==1.7
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
cut-cross-entropy==25.1.1
cycler==0.12.1
Cython==3.1.2
dashscope==1.24.1
dataclasses-json==0.6.7
datasets==3.5.0
decorator==5.2.1
diffusers==0.34.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
durationpy==0.10
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.11.0.post1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
GPUtil==1.4.0
gradio==5.25.2
gradio_client==1.8.0
graphviz==0.20.3
greenlet==3.2.3
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
hf_transfer==0.1.9
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental==0.3.4
langchain-ollama==0.3.6
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langchain-unstructured==0.1.6
langdetect==1.0.9
langsmith==0.4.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lmdb==1.6.2
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.6.0
mdurl==0.1.2
mkl==2021.4.0
mmh3==5.2.0
modelscope==1.29.0
modelscope_studio==1.4.4
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.3.2
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numexpr==2.11.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nvjitlink-cu12==12.9.86
oauthlib==3.3.1
olefile==0.47
ollama==0.5.1
onnxruntime==1.19.2
open_clip_torch @ git+https://gitee.com/ufhy/open_clip.git@bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b
openai==1.93.0
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt-einsum==3.3.0
orjson==3.10.16
oss2==2.19.1
overrides==7.7.0
packaging==24.2
paddleocr==3.1.0
paddlepaddle-gpu==3.1.0
paddlex==3.1.1
pandas==2.2.3
peft==0.12.0
pillow==11.1.0
platformdirs==4.3.8
pooch==1.8.2
posthog==5.4.0
premailer==3.10.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.8.0
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
qianfan==0.4.12.3
qwen-omni-utils==0.0.8
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
shtab==1.7.2
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spaces==0.39.0
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.44.1
stringzilla==3.12.5
sympy==1.14.0
tbb==2021.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchsummary==1.5.1
torchvision==0.22.1+cu126
tornado==6.4.2
tqdm==4.67.1
transformers==4.55.2
triton-windows==3.3.1.post19
trl==0.19.0
typeguard==4.4.4
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.14.0
tyro==0.9.26
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.180
ultralytics-thop==2.0.15
unsloth_zoo==2025.6.8
unstructured==0.18.11
unstructured-client==0.41.0
urllib3==2.3.0
uvicorn==0.34.1
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whisper==1.1.10
wikipedia==1.4.0
wrapt==1.17.2
xformers==0.0.31.post1
xxhash==3.5.0
yarl==1.19.0
zai-sdk==0.0.3
zhipuai==2.1.5.20250801
zipp==3.23.0
zstandard==0.23.0
